合并

+++ src/think/console/command/Clear.php
@@ -12,7 +12,6 @@ namespace think\console\command;
 
 use think\console\Command;
 use think\console\Input;
-use think\console\input\Argument;
 use think\console\input\Option;
 use think\console\Output;
 
@@ -21,56 +20,66 @@ class Clear extends Command
     protected function configure()
     {
         // 指令配置
-        $this
-            ->setName('clear')
-            ->addArgument('app', Argument::OPTIONAL, 'Build app config cache .')
+        $this->setName('clear')
             ->addOption('path', 'd', Option::VALUE_OPTIONAL, 'path to clear', null)
             ->addOption('cache', 'c', Option::VALUE_NONE, 'clear cache file')
             ->addOption('log', 'l', Option::VALUE_NONE, 'clear log file')
             ->addOption('dir', 'r', Option::VALUE_NONE, 'clear empty dir')
-            ->addOption('route', 'u', Option::VALUE_NONE, 'clear route cache')
+            ->addOption('expire', 'e', Option::VALUE_NONE, 'clear cache file if cache has expired')
             ->setDescription('Clear runtime file');
     }
 
     protected function execute(Input $input, Output $output)
     {
-        if ($input->getOption('route')) {
-            $this->app->cache->clear('route_cache');
+        $runtimePath = $this->app->getRootPath() . 'runtime' . DIRECTORY_SEPARATOR;
+
+        if ($input->getOption('cache')) {
+            $path = $runtimePath . 'cache';
+        } elseif ($input->getOption('log')) {
+            $path = $runtimePath . 'log';
         } else {
-            if ($input->getArgument('app')) {
-                $runtimePath = $this->app->getRootPath() . 'runtime' . DIRECTORY_SEPARATOR . $input->getArgument('app') . DIRECTORY_SEPARATOR;
-            } else {
-                $runtimePath = $this->app->getRuntimePath();
-            }
-
-            if ($input->getOption('cache')) {
-                $path = $runtimePath . 'cache';
-            } elseif ($input->getOption('log')) {
-                $path = $runtimePath . 'log';
-            } else {
-                $path = $input->getOption('path') ?: $runtimePath;
-            }
-
-            $rmdir = $input->getOption('dir') ? true : false;
-            $this->clear(rtrim($path, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR, $rmdir);
+            $path = $input->getOption('path') ?: $runtimePath;
         }
+
+        $rmdir = $input->getOption('dir') ? true : false;
+        // --expire 仅当 --cache 时生效
+        $cache_expire = $input->getOption('expire') && $input->getOption('cache') ? true : false;
+        $this->clear(rtrim($path, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR, $rmdir, $cache_expire);
 
         $output->writeln("<info>Clear Successed</info>");
     }
 
-    protected function clear(string $path, bool $rmdir): void
+    protected function clear(string $path, bool $rmdir, bool $cache_expire): void
     {
         $files = is_dir($path) ? scandir($path) : [];
 
         foreach ($files as $file) {
             if ('.' != $file && '..' != $file && is_dir($path . $file)) {
-                array_map('unlink', glob($path . $file . DIRECTORY_SEPARATOR . '*.*'));
+                $this->clear($path . $file . DIRECTORY_SEPARATOR, $rmdir, $cache_expire);
                 if ($rmdir) {
-                    rmdir($path . $file);
+                    @rmdir($path . $file);
                 }
             } elseif ('.gitignore' != $file && is_file($path . $file)) {
-                unlink($path . $file);
+                if ($cache_expire) {
+                    if ($this->cacheHasExpired($path . $file)) {
+                        unlink($path . $file);
+                    }
+                } else {
+                    unlink($path . $file);
+                }
             }
         }
     }
+
+    /**
+     * 缓存文件是否已过期
+     * @param $filename string 文件路径
+     * @return bool
+     */
+    protected function cacheHasExpired($filename) {
+        $content = file_get_contents($filename);
+        $expire = (int) substr($content, 8, 12);
+        return 0 != $expire && time() - $expire > filemtime($filename);
+    }
+
 }
